requirements.txt
Flask
pandas
pillow
aiohttp
mysql-connector-python